all_requirements.txt
transformers>=4.34.0
Pillow
python-Levenshtein
imagesize
accelerate
tensorboard
natsort
setproctitle
anyconfig
hypercorn
flask